requirements.txt
atproto==0.0.55
websockets==12.0
requests==2.31.0
python-dateutil==2.9.0
python-dotenv==1.0.0
sentence-transformers==3.0.1
scikit-learn==1.5.1
numpy==1.26.4
torch==2.3.1
redis==5.0.1
fastapi==0.104.1
uvicorn==0.24.0